refactor dedup_test.go
